Changed movement calculation. Fixed FPS based gameloop. Added a child object that places itself infront of its owner. CHanged how audio is played. Added trigger colliders. Fixed rendering.

+++ game_engine/src/game_engine/Timer.java
@@ -48,7 +48,7 @@ public class Timer {
 		return delta;
 	}
 	
-	public float getRunTime () {
+	public long getRunTime () {
 		long runTime = ((System.currentTimeMillis() - startTime));
 		return runTime;
 	}
